+++ source/renderer/app/components/loading/syncing-connecting/SyncingConnectingBackground.js
@@ -5,13 +5,12 @@ import classNames from 'classnames';
 import SVGInline from 'react-svg-inline';
 import backgroundImage from '../../../assets/images/circle-bg-faded.inline.svg';
 import styles from './SyncingConnectingBackground.scss';
-import { THEMES } from '../../../themes';
 
 type Props = {
-  currentTheme: string,
+  hasLoadedCurrentTheme: boolean,
+  isIncentivizedTestnet: boolean,
   isConnecting: boolean,
   isSyncing: boolean,
-  hasLoadedCurrentTheme: boolean,
 };
 
 @observer
@@ -20,18 +19,18 @@ export default class SyncingConnectingBackground extends Component<Props> {
     const {
       isConnecting,
       isSyncing,
+      isIncentivizedTestnet,
       hasLoadedCurrentTheme,
-      currentTheme,
     } = this.props;
     const componentStyles = classNames([
       styles.component,
-      hasLoadedCurrentTheme ? null : styles.isLoadingTheme,
+      !hasLoadedCurrentTheme ? styles.isLoadingTheme : null,
       isConnecting ? styles.isConnecting : null,
       isSyncing ? styles.isSyncing : null,
     ]);
     return (
       <div className={componentStyles}>
-        {currentTheme === THEMES.INCENTIVIZED_TESTNET && (
+        {isIncentivizedTestnet && (
           <>
             <div className={styles.backgroundOverlay} />
             <SVGInline