+++ source/renderer/app/components/splash/Network.js
@@ -8,7 +8,6 @@ import backgroundImage from '../../assets/images/circle-bg-faded.inline.svg';
 import daedalusIcon from '../../assets/images/daedalus-logo-loading-grey.inline.svg';
 import externalLinkIcon from '../../assets/images/link-ic.inline.svg';
 import styles from './Network.scss';
-import { THEMES } from '../../themes';
 
 const messages = defineMessages({
   title: {
@@ -46,7 +45,7 @@ const messages = defineMessages({
 });
 
 type Props = {
-  currentTheme: string,
+  isIncentivizedTestnet: boolean,
   onClose: Function,
   onLearnMoreClick: Function,
 };
@@ -58,7 +57,7 @@ export default class SplashNetwork extends Component<Props> {
 
   render() {
     const { intl } = this.context;
-    const { onClose, onLearnMoreClick, currentTheme } = this.props;
+    const { isIncentivizedTestnet, onClose, onLearnMoreClick } = this.props;
     const title = intl.formatMessage(messages.title);
     const subTitle1 = intl.formatMessage(messages.incentivizedTestnet);
     const subTitle2 = intl.formatMessage(messages.balanceCheck);
@@ -70,7 +69,7 @@ export default class SplashNetwork extends Component<Props> {
     return (
       <div className={styles.component}>
         <div className={styles.backgroundContainer}>
-          {currentTheme === THEMES.INCENTIVIZED_TESTNET && (
+          {isIncentivizedTestnet && (
             <>
               <div className={styles.backgroundOverlay} />
               <SVGInline
@@ -82,9 +81,13 @@ export default class SplashNetwork extends Component<Props> {
         </div>
         <div className={styles.content}>
           <SVGInline svg={daedalusIcon} className={styles.daedalusIcon} />
-          <div className={styles.title}>{title}</div>
-          <div className={styles.subTitle1}>{subTitle1}</div>
-          <div className={styles.subTitle2}>{subTitle2}</div>
+          {isIncentivizedTestnet && (
+            <>
+              <div className={styles.title}>{title}</div>
+              <div className={styles.subTitle1}>{subTitle1}</div>
+              <div className={styles.subTitle2}>{subTitle2}</div>
+            </>
+          )}
           <div className={styles.description}>{description}</div>
           <div className={styles.action}>
             <Button