+++ storybook/stories/navigation/Sidebar.stories.js
@@ -63,14 +63,11 @@ const sidebarMenus = observable({
 
 let emptyMenus;
 
-let currentTheme = sessionStorage.getItem('themeName') || 'light-blue';
-currentTheme = currentTheme.toLowerCase();
-
 storiesOf('Navigation|Sidebar', module)
   .addDecorator(story => <StoryDecorator>{story()}</StoryDecorator>)
   .addDecorator(withKnobs)
   // ====== Stories ======
-  .add('No Category', () => (
+  .add('No Category', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -81,12 +78,12 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
       isIncentivizedTestnet
     />
   ))
-  .add('Wallets Category', () => (
+  .add('Wallets Category', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -97,12 +94,12 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
       isIncentivizedTestnet
     />
   ))
-  .add('Wallet Selected', () => (
+  .add('Wallet Selected', (props: { currentTheme: string }) => (
     <Sidebar
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
       activeSidebarCategory={CATEGORIES_WITH_DELEGATION_COUNTDOWN[0].route}
@@ -114,12 +111,12 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
       isIncentivizedTestnet
     />
   ))
-  .add('Delegation Category', () => (
+  .add('Delegation Category', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -130,28 +127,31 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
       isIncentivizedTestnet
     />
   ))
-  .add('Decentralization Progress Category', () => (
-    <Sidebar
-      menus={emptyMenus}
-      categories={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN}
-      activeSidebarCategory={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN[1].route}
-      onActivateCategory={action('onActivateCategory')}
-      isDialogOpen={() => false}
-      onAddWallet={action('onAddWallet')}
-      onOpenDialog={action('openDialog')}
-      onSubmitSupportRequest={() => {}}
-      pathname="/"
-      currentTheme={currentTheme}
-      network="testnet"
-      isIncentivizedTestnet
-    />
-  ))
-  .add('Network label', () => (
+  .add(
+    'Decentralization Progress Category',
+    (props: { currentTheme: string }) => (
+      <Sidebar
+        menus={emptyMenus}
+        categories={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN}
+        activeSidebarCategory={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN[1].route}
+        onActivateCategory={action('onActivateCategory')}
+        isDialogOpen={() => false}
+        onAddWallet={action('onAddWallet')}
+        onOpenDialog={action('openDialog')}
+        onSubmitSupportRequest={() => {}}
+        pathname="/"
+        currentTheme={props.currentTheme}
+        network="testnet"
+        isIncentivizedTestnet
+      />
+    )
+  )
+  .add('Network label', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -162,7 +162,7 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network={select(
         'Netork badge',
         {